Reorg setbfree ui to accommodate reverb mix
This change splits the vibrato & percussion screen to their own
screens and moves their on/off controls to those screens.

This allows the reverb mix to be added to the main screen.

The main, vibrato and percussion screens now only have 3 controls
assigned, so there is potential to add more in a future change.