setbfree rotary use speed-select and speed-toggle
The rotary speed-preset only has three settings and doesn't allow the
speed of the horn and drum to be set separately.

Using speed-select instead allows all combinations of off/slow/fast
and horn/drum to be selected.

Also, having the rotary toggle (sustain pedal) be visible in the ui is
important, since it is very common to hear the rotary go from off to
the selected speed.