Map presets to pedalboard presets
Currently the zynthian presets are mapped to mod-ui effect presets, however
this mapping doesn't work very naturally in practice. It generally ends up with
a preset list where every preset is called 'Default' and it is not clear that
each preset might be for a different effect.

Fortunately mod-ui also has the concept of a pedalboard preset (possible this
didn't exist when the engine was first written). A pedalboard preset maps very
well to a zynthian preset, so this change replaces the effect preset with a
pedalboard preset implementation.